Make wkhtmltopdf output using --print-media-type
Without this option page breaks do not appear in the output PDF using the
default template. The reason for this is that CSS rules scoped with "@media
print" are not applied unless wkhtmltopdf uses the print media type.

An alternate solution to this problem would be to unscope the CSS rules.